First not abviously wrong version of single-pass top.
Notice that error ranges are very bad when key is the one used to hash
values amongst various hdirs since we then merge non-random datasets.
That's much better for random keys (like mac addrs) for wich 0-error can
be achieved.

When using ip-srv in key, we should use the knownledge that keys that are
not hashed the same cannot be found in a dataset... :-/